middleware + cleanup
