+++ router/middleware.go
@@ -1,0 +1,27 @@
+package router
+
+import (
+	"fmt"
+	"net/http"
+	"time"
+)
+
+func logMiddleware(handler http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		nw := time.Now()
+		handler.ServeHTTP(w, r)
+		fmt.Printf("%s %s %s %s\n", r.RemoteAddr, r.Method, r.URL, time.Since(nw))
+	})
+}
+
+func recoverMiddleware(handler http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		defer func() {
+			if err := recover(); err != nil {
+				fmt.Printf("RECOVERING: %v\n", err)
+				http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+			}
+		}()
+		handler.ServeHTTP(w, r)
+	})
+}
