+++ router/router.go
@@ -11,9 +11,8 @@ import (
 func NewRouter(initTasks bool) http.Handler {
 	mux := http.NewServeMux()
 
-	mux.HandleFunc("/", taskListPage())
-
 	th := NewTasksHandler(initTasks)
+	mux.HandleFunc("/", th.taskListPage())
 	mux.HandleFunc("/new", th.taskCreatePage())
 	mux.Handle("/view/", http.StripPrefix("/view", http.HandlerFunc(th.taskViewPage())))
 	mux.Handle("/edit/", http.StripPrefix("/edit", http.HandlerFunc(th.taskEditPage())))
@@ -22,26 +21,7 @@ func NewRouter(initTasks bool) http.Handler {
 	mux.HandleFunc("/api/comments", th.addComment)
 	mux.Handle("/static/", http.StripPrefix("/static/", http.FileServer(http.Dir("./web/static"))))
 
-	return mux
-}
-
-func taskListPage() http.HandlerFunc {
-	files := tmplLayout("./web/templates/index.gohtml")
-	tmpl := template.Must(template.New("index").Funcs(defaultFuncs).ParseFiles(files...))
-	return func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/" || r.Method != http.MethodGet {
-			notFoundPage(w, r)
-			return
-		}
-		var buf bytes.Buffer
-		if err := tmpl.ExecuteTemplate(&buf, "base", nil); err != nil {
-			fmt.Printf("ERR: %v\n", err)
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-		io.Copy(w, &buf)
-	}
+	return logMiddleware(recoverMiddleware(mux))
 }
 
 var notFoundPage = func() http.HandlerFunc {
